Issue-8: See courses not yet taken
I have created methods to get a course object given a course name or id as a parameter. I have also added methods that let you determine if you can take a course based on if that student has that prerequisite. Also, you can get a list of all courses that the student can take that they have not taken based on whether or not they have those prerequisites
Spent: 2 hours