Created unit tests for objects, added getters for private variables
Created new test files for each business object, each including a test to see that the data is correctly stored in the object being tested.
Added getters as needed in order to pass the tests.
Added comments to each class file in the objects package, briefly describing the purpose of the class.
Changed CourseType to DegreeCourseType, functionality for the renamed class is unchanged.